+++ e2e/test-env/src/client/RicosTestApp/RicosTestApp.jsx
@@ -12,35 +12,38 @@ import theme from '../../../../../examples/main/shared/theme/theme';
 
 class RicosTestApp extends PureComponent {
   renderEditor = () => {
-    const toolbarsConfig = {
-      addPluginMenuConfig: {
-        showSearch: true,
-        splitToSections: true,
-      },
-    };
+    // const addPluginMenuConfig = {
+    //   showSearch: true,
+    //   splitToSections: true,
+    // };
+    // const toolbarSettings = {
+    //   getToolbarSettings: () => [
+    //     { name: 'SIDE', addPluginMenuConfig },
+    //     { name: 'MOBILE', addPluginMenuConfig },
+    //   ],
+    // };
 
-    const { contentState, onEditorChange, locale, isMobile } = this.props;
+    const { contentState, onEditorChange, locale, isMobile, testAppConfig } = this.props;
     return (
       <RicosEditor
-        plugins={editorPlugins}
+        plugins={editorPlugins(testAppConfig.plugins)}
         placeholder={'Add some text!'}
         content={contentState}
         isMobile={isMobile}
         locale={locale}
         cssOverride={theme}
-        _rcProps={toolbarsConfig} // DO NOT use _rcProps for any other prop
+        //toolbarSettings={toolbarSettings}
       >
-        <RichContentEditor onChange={onEditorChange} />
+        <RichContentEditor onChange={onEditorChange} config={testAppConfig.pluginsConfig} />
       </RicosEditor>
     );
   };
 
   renderViewer = () => {
-    const { isMobile, contentState, locale, seoMode } = this.props;
-
+    const { isMobile, contentState, locale, seoMode, testAppConfig } = this.props;
     return (
       <RicosViewer
-        plugins={viewerPlugins}
+        plugins={viewerPlugins(testAppConfig.plugins)}
         content={contentState}
         isMobile={isMobile}
         locale={locale}
@@ -76,6 +79,7 @@ RicosTestApp.propTypes = {
   localeResource: PropTypes.object,
   onEditorChange: PropTypes.func,
   seoMode: PropTypes.bool,
+  testAppConfig: PropTypes.object,
 };
 
 export default windowContentStateHoc(RicosTestApp);