+++ e2e/test-env/src/client/RicosTestApp/editorPlugins.js
@@ -1,4 +1,4 @@
-import { pluginLinkButton } from 'wix-rich-content-plugin-button';
+import { pluginLinkButton, pluginActionButton } from 'wix-rich-content-plugin-button';
 import { pluginCodeBlock } from 'wix-rich-content-plugin-code-block';
 import { pluginDivider } from 'wix-rich-content-plugin-divider';
 import { pluginEmoji } from 'wix-rich-content-plugin-emoji';
@@ -6,25 +6,39 @@ import { pluginFileUpload } from 'wix-rich-content-plugin-file-upload';
 import { pluginGallery } from 'wix-rich-content-plugin-gallery';
 import { pluginGiphy } from 'wix-rich-content-plugin-giphy';
 import { pluginHashtag } from 'wix-rich-content-plugin-hashtag';
+import { pluginHeadings } from 'wix-rich-content-plugin-headings';
 import { pluginHeadersMarkdown } from 'wix-rich-content-plugin-headers-markdown';
 import { pluginHtml } from 'wix-rich-content-plugin-html';
 import { pluginImage } from 'wix-rich-content-plugin-image';
 import { pluginLineSpacing } from 'wix-rich-content-plugin-line-spacing';
+import { pluginIndent } from 'wix-rich-content-plugin-indent';
 import { pluginLink } from 'wix-rich-content-plugin-link';
 import { pluginMap } from 'wix-rich-content-plugin-map';
 import { pluginMentions } from 'wix-rich-content-plugin-mentions';
 import { pluginSoundCloud } from 'wix-rich-content-plugin-sound-cloud';
 import { pluginUndoRedo } from 'wix-rich-content-plugin-undo-redo';
 import { pluginVideo } from 'wix-rich-content-plugin-video';
-import { pluginLinkPreview } from 'wix-rich-content-plugin-link-preview';
+import { pluginLinkPreview, LinkPreviewProviders } from 'wix-rich-content-plugin-link-preview';
+import {
+  pluginVerticalEmbed,
+  verticalEmbedProviders,
+} from 'wix-rich-content-plugin-vertical-embed';
 import { mockFetchUrlPreviewData } from '../../../../../examples/main/shared/utils/linkPreviewUtil';
 import { pluginTextColor, pluginTextHighlight } from 'wix-rich-content-plugin-text-color';
+
+import { createPresets } from './utils';
 import {
   customForegroundStyleFn,
   styleSelectionPredicate,
   colorScheme,
   customBackgroundStyleFn,
 } from '../../../../../examples/main/src/text-color-style-fn';
+
+// eslint-disable-next-line max-len
+import MockVerticalSearchModule from '../../../../../examples/main/shared/editor/Utils/verticalEmbedUtil';
+
+const { Instagram, Twitter, YouTube, TikTok } = LinkPreviewProviders;
+const { product } = verticalEmbedProviders;
 
 const configs = {
   fileUpload: {
@@ -54,38 +68,59 @@ const configs = {
   linkPreview: {
     fetchData: mockFetchUrlPreviewData(),
     enableEmbed: true,
+    exposeEmbedButtons: [Instagram, Twitter, YouTube, TikTok],
   },
-};
-const plugins = [
-  pluginImage({ handleFileSelection: () => true }),
-  pluginGallery({ handleFileSelection: () => true }),
-  pluginVideo(),
-  pluginHtml(),
-  pluginDivider(),
-  pluginCodeBlock(),
-  pluginSoundCloud(),
-  pluginGiphy(configs.giphy),
-  pluginMap({ googleMapApiKey: process.env.GOOGLE_MAPS_API_KEY }),
-  pluginFileUpload(configs.fileUpload),
-  pluginLinkButton(),
-  pluginEmoji(),
-  pluginHashtag(),
-  pluginHeadersMarkdown(),
-  pluginLineSpacing(),
-  pluginLink(),
-  pluginMentions(),
-  pluginLinkPreview(configs.linkPreview),
-  pluginUndoRedo(),
-  pluginTextHighlight({
+  verticalEmbed: {
+    verticalsApi: type => new MockVerticalSearchModule(type),
+    // exposeEmbedButtons: [product, event, booking],
+    exposeEmbedButtons: [product],
+  },
+  textHighlight: {
     colorScheme,
     styleSelectionPredicate,
     customStyleFn: customBackgroundStyleFn,
-  }),
-  pluginTextColor({
+  },
+  textColor: {
     colorScheme,
     styleSelectionPredicate,
     customStyleFn: customForegroundStyleFn,
-  }),
-];
+  },
+};
 
-export default plugins;
+const plugins = {
+  image: pluginImage({ handleFileSelection: () => true }),
+  gallery: pluginGallery({ handleFileSelection: () => true }),
+  video: pluginVideo(),
+  html: pluginHtml(),
+  divider: pluginDivider(),
+  codeBlock: pluginCodeBlock(),
+  link: pluginLink(),
+  linkPreview: pluginLinkPreview(configs.linkPreview),
+  spacing: pluginLineSpacing(),
+  indent: pluginIndent(),
+  hashtag: pluginHashtag(),
+  mentions: pluginMentions(),
+  soundCloud: pluginSoundCloud(),
+  giphy: pluginGiphy(configs.giphy),
+  headers: pluginHeadersMarkdown(),
+  map: pluginMap({ googleMapApiKey: process.env.GOOGLE_MAPS_API_KEY }),
+  fileUpload: pluginFileUpload(configs.fileUpload),
+  linkButton: pluginLinkButton(),
+  actionButton: pluginActionButton(),
+  highlight: pluginTextHighlight(configs.textHighlight),
+  textColor: pluginTextColor(configs.textColor),
+  emoji: pluginEmoji(),
+  undoRedo: pluginUndoRedo(),
+  headings: pluginHeadings(),
+  verticalEmbed: pluginVerticalEmbed(configs.verticalEmbed),
+};
+
+const presets = createPresets(plugins);
+
+export default pluginsPreset =>
+  pluginsPreset
+    ? pluginsPreset
+        .map(plugin => presets[plugin])
+        .flat()
+        .filter(val => !!val)
+    : presets.all;