+++ e2e/test-env/src/client/RicosTestApp/viewerPlugins.js
@@ -1,4 +1,7 @@
-import { pluginLinkButton } from 'wix-rich-content-plugin-button/dist/module.viewer';
+import {
+  pluginLinkButton,
+  pluginActionButton,
+} from 'wix-rich-content-plugin-button/dist/module.viewer';
 import { pluginCodeBlock } from 'wix-rich-content-plugin-code-block/dist/module.viewer';
 import { pluginDivider } from 'wix-rich-content-plugin-divider/dist/module.viewer';
 import { pluginEmoji } from 'wix-rich-content-plugin-emoji/dist/module.viewer';
@@ -6,6 +9,7 @@ import { pluginFileUpload } from 'wix-rich-content-plugin-file-upload/dist/modul
 import { pluginGallery } from 'wix-rich-content-plugin-gallery/dist/module.viewer';
 import { pluginGiphy } from 'wix-rich-content-plugin-giphy/dist/module.viewer';
 import { pluginHashtag } from 'wix-rich-content-plugin-hashtag/dist/module.viewer';
+import { pluginIndent } from 'wix-rich-content-plugin-indent';
 import { pluginHeadersMarkdown } from 'wix-rich-content-plugin-headers-markdown/dist/module.viewer';
 import { pluginHtml } from 'wix-rich-content-plugin-html/dist/module.viewer';
 import { pluginImage } from 'wix-rich-content-plugin-image/dist/module.viewer';
@@ -16,10 +20,12 @@ import { pluginMentions } from 'wix-rich-content-plugin-mentions/dist/module.vie
 import { pluginSoundCloud } from 'wix-rich-content-plugin-sound-cloud/dist/module.viewer';
 import { pluginVideo } from 'wix-rich-content-plugin-video/dist/module.viewer';
 import { pluginLinkPreview } from 'wix-rich-content-plugin-link-preview/dist/module.viewer';
+import { pluginVerticalEmbed } from 'wix-rich-content-plugin-vertical-embed/dist/module.viewer';
 import {
   pluginTextColor,
   pluginTextHighlight,
 } from 'wix-rich-content-plugin-text-color/dist/module.viewer';
+import { createPresets } from './utils';
 import {
   viewerCustomForegroundStyleFn,
   styleSelectionPredicate,
@@ -51,35 +57,51 @@ const configs = {
     giphySdkApiKey: process.env.GIPHY_API_KEY || 'HXSsAGVNzjeUjhKfhhD9noF8sIbpYDsV',
     sizes: { desktop: 'original', mobile: 'original' }, // original or downsizedSmall are supported
   },
-};
-
-const plugins = [
-  pluginLinkButton(),
-  pluginCodeBlock(),
-  pluginDivider(),
-  pluginEmoji(),
-  pluginFileUpload(configs.fileUpload),
-  pluginGallery(),
-  pluginGiphy(configs.giphy),
-  pluginHashtag(),
-  pluginHtml(),
-  pluginImage(),
-  pluginHeadersMarkdown(),
-  pluginLineSpacing(),
-  pluginLink(),
-  pluginMap({ googleMapApiKey: process.env.GOOGLE_MAPS_API_KEY }),
-  pluginMentions(),
-  pluginSoundCloud(),
-  pluginVideo(),
-  pluginTextHighlight({
+  linkPreview: {
+    enableEmbed: true,
+  },
+  textHighlight: {
     styleSelectionPredicate,
     customStyleFn: viewerCustomBackgroundStyleFn,
-  }),
-  pluginTextColor({
+  },
+  textColor: {
     styleSelectionPredicate,
     customStyleFn: viewerCustomForegroundStyleFn,
-  }),
-  pluginLinkPreview({ enableEmbed: true }),
-];
+  },
+};
 
-export default plugins;
+const plugins = {
+  image: pluginImage({ handleFileSelection: () => true }),
+  gallery: pluginGallery({ handleFileSelection: () => true }),
+  video: pluginVideo(),
+  html: pluginHtml(),
+  divider: pluginDivider(),
+  spacing: pluginLineSpacing(),
+  link: pluginLink(),
+  linkPreview: pluginLinkPreview(configs.linkPreview),
+  indent: pluginIndent(),
+  hashtag: pluginHashtag(),
+  mentions: pluginMentions(),
+  codeBlock: pluginCodeBlock(),
+  soundCloud: pluginSoundCloud(),
+  giphy: pluginGiphy(configs.giphy),
+  headers: pluginHeadersMarkdown(),
+  map: pluginMap({ googleMapApiKey: process.env.GOOGLE_MAPS_API_KEY }),
+  fileUpload: pluginFileUpload(configs.fileUpload),
+  linkButton: pluginLinkButton(),
+  actionButton: pluginActionButton(),
+  highlight: pluginTextHighlight(configs.textHighlight),
+  textColor: pluginTextColor(configs.textColor),
+  emoji: pluginEmoji(),
+  verticalEmbed: pluginVerticalEmbed(),
+};
+
+const presets = createPresets(plugins);
+
+export default pluginsPreset =>
+  pluginsPreset
+    ? pluginsPreset
+        .map(plugin => presets[plugin])
+        .flat()
+        .filter(val => !!val)
+    : presets.all;